SemiDynamic Navigation Meshes
Imported NavMeshComponents (https://github.com/Unity-Technologies/NavMeshComponents)

Integrated NavMesh generation with maze generation, which creates an AI-navigable maze at runtime.

I don't know how process intensive rebuilding this mesh is (it needs to be built twice during generation but this can be hidden behind a loading screen), so it may effect the whole idea of having the maze regenerate itself.

It's also unlikely that this navmesh could accurately describe non-euclidean areas of the maze, but that's a sacrifice I'm willing to make.