Basic Nclidian Maze Cells and Helper Data Structures
Added scripts and prefabs for the Nclidian sections, which use Lague's portals as a base. The Portals themselves are replacers that are managed by an NclidianController until they are placed. Also created a MazeNeighbors component for easy management of a maze's neighbors. It'll be used for connecting portals to the maze after generation.

Also added a script and prefab for a dynamic spawn room. Probably not going to use it, but it's there.

Also updated code license notices to be more descriptive, and thus safer. I know it should have probably been a separate commits but monkey brain demanded updating it immediately. Oh well, it happens.