Restored Portal Material Updating
The method that establishes which portals need to be rendered simply needed to be called again, since the portals don't exist at runtime.

Currently this happens via MainCameraUpdatable, which is just Lague's MainCamera script, but seperated and with a message-callable method. Might be moved to the PlayerCharacter script later for consistency.