Refactored MazeGenerator to place the actual maze structure in its own class
MazeGenerator is no longer a MonoBehavior. Rather, Maze is attached to gameobjects to represent mazes. Each maze comes with a MazeGenerator object though, so a Maze's default generation settings can still be applied in the inspector. However, a different settings object can be used...

Also disabled most of SpawnRoom, which was depreciated by this refactoring