Added functionality to freeze arbitrary gameObjects when dialogue starts
Specifically, when dialogue starts (or stops), the slow FindObjectsOfType<>() method runs on the type FreezeableNPC

FreezableNPC is an abstract class whose implemented methods contain the code that does the actual freezing or unfreezing.

Currently FreezablePlayer can freeze the player (and it is used) and FreezableNavAgent can freeze NavMeshAgents (but it isn't used yet).