Added Vault Controllers to introduce POIs to the Maze
Currently, the portals that lead to them do not function 100% of the time, but they do generate correctly in the maze and they still at least work sometimes.

Problems: Floors between Vault Portals are uneven and the player occasionally comes out of the wrong side of the vault-side portal.

These issues are most likely due to one of the portals being used raw, rather than replacing a true maze cell.