Player Death Part 1
When the player is touched by the minotaur, the player will rotate towards the minotaur (and vice versa), and then the game will switch to a death scene.

The death scene doesn't exist yet, and the rotation is weird because the player's camera is higher (not lower) than the minotaur's position (which is its feet), but everything else works so far.