Player Death Part 2
The player will now die to a simple death scene which offers options to restart or end the game.

Also gutted Minotaur's pathing. It now won't adapt to player movement, but it won't freeze up anymore either.

Also tried a new algo for getting the player to rotate towards the mino's head. It didn't work, I'll think of a new one.